rename SetPropertyRouteHelper to SetPropertyBase and convert to a base class for both setter routes
