+++ web/web-base/src/main/java/org/visallo/web/routes/SetPropertyBase.java
@@ -1,6 +1,5 @@
-package org.visallo.web;
+package org.visallo.web.routes;
 
-import com.google.inject.Inject;
 import org.vertexium.Authorizations;
 import org.vertexium.Graph;
 import org.visallo.core.exception.VisalloException;
@@ -9,6 +8,7 @@ import org.visallo.core.security.VisibilityTranslator;
 import org.visallo.core.user.User;
 import org.visallo.core.util.VisalloLogger;
 import org.visallo.core.util.VisalloLoggerFactory;
+import org.visallo.web.BadRequestException;
 
 import javax.servlet.http.HttpServletRequest;
 import java.text.SimpleDateFormat;
@@ -16,27 +16,26 @@ import java.util.Date;
 import java.util.ResourceBundle;
 import java.util.TimeZone;
 
-public class SetPropertyRouteHelper {
-    private static final VisalloLogger LOGGER = VisalloLoggerFactory.getLogger(SetPropertyRouteHelper.class);
+public abstract class SetPropertyBase {
+    private static final VisalloLogger LOGGER = VisalloLoggerFactory.getLogger(SetPropertyBase.class);
 
-    private Graph graph;
-    private VisibilityTranslator visibilityTranslator;
+    protected final Graph graph;
+    protected final VisibilityTranslator visibilityTranslator;
 
-    @Inject
-    public SetPropertyRouteHelper(Graph graph, VisibilityTranslator visibilityTranslator) {
+    protected SetPropertyBase(Graph graph, VisibilityTranslator visibilityTranslator) {
         this.graph = graph;
         this.visibilityTranslator = visibilityTranslator;
     }
 
-    public boolean isCommentProperty(String propertyName) {
+    protected boolean isCommentProperty(String propertyName) {
         return VisalloProperties.COMMENT.isSameName(propertyName);
     }
 
-    public String createPropertyKey(String propertyName, Graph graph) {
+    protected String createPropertyKey(String propertyName, Graph graph) {
         return isCommentProperty(propertyName) ? createCommentPropertyKey() : graph.getIdGenerator().nextId();
     }
 
-    public void checkVisibilityParameter(
+    protected void checkVisibilityParameter(
             String visibilitySource, Authorizations authorizations, User user, ResourceBundle resourceBundle) {
         if (!graph.isVisibilityValid(visibilityTranslator.toVisibility(visibilitySource).getVisibility(), authorizations)) {
             LOGGER.warn("%s is not a valid visibility for %s user", visibilitySource, user.getDisplayName());
@@ -44,7 +43,7 @@ public class SetPropertyRouteHelper {
         }
     }
 
-    public void checkRoutePath(String entityType, String propertyName, HttpServletRequest request) {
+    protected void checkRoutePath(String entityType, String propertyName, HttpServletRequest request) {
         boolean isComment = isCommentProperty(propertyName);
         if (isComment && request.getPathInfo().equals(String.format("/%s/property", entityType))) {
             throw new VisalloException(String.format("Use /%s/comment to save comment properties", entityType));
