+++ web/web-base/src/main/java/org/visallo/web/routes/edge/EdgeSetProperty.java
@@ -21,7 +21,7 @@ import org.visallo.core.util.VertexiumMetadataUtil;
 import org.visallo.core.util.VisalloLogger;
 import org.visallo.core.util.VisalloLoggerFactory;
 import org.visallo.web.BadRequestException;
-import org.visallo.web.SetPropertyRouteHelper;
+import org.visallo.web.routes.SetPropertyBase;
 import org.visallo.web.VisalloResponse;
 import org.visallo.web.clientapi.model.ClientApiSourceInfo;
 import org.visallo.web.clientapi.model.ClientApiSuccess;
@@ -31,17 +31,14 @@ import org.visallo.web.parameterProviders.JustificationText;
 import javax.servlet.http.HttpServletRequest;
 import java.util.ResourceBundle;
 
-public class EdgeSetProperty implements ParameterizedHandler {
+public class EdgeSetProperty extends SetPropertyBase implements ParameterizedHandler {
     private static final VisalloLogger LOGGER = VisalloLoggerFactory.getLogger(EdgeSetProperty.class);
 
-    private final Graph graph;
     private final OntologyRepository ontologyRepository;
-    private VisibilityTranslator visibilityTranslator;
     private final WorkQueueRepository workQueueRepository;
     private final WorkspaceRepository workspaceRepository;
     private final GraphRepository graphRepository;
     private final ACLProvider aclProvider;
-    private final SetPropertyRouteHelper routeHelper;
     private final boolean autoPublishComments;
 
     @Inject
@@ -53,17 +50,14 @@ public class EdgeSetProperty implements ParameterizedHandler {
             final WorkspaceRepository workspaceRepository,
             final GraphRepository graphRepository,
             final ACLProvider aclProvider,
-            final SetPropertyRouteHelper routeHelper,
             final Configuration configuration
     ) {
+        super(graph, visibilityTranslator);
         this.ontologyRepository = ontologyRepository;
-        this.graph = graph;
-        this.visibilityTranslator = visibilityTranslator;
         this.workQueueRepository = workQueueRepository;
         this.workspaceRepository = workspaceRepository;
         this.graphRepository = graphRepository;
         this.aclProvider = aclProvider;
-        this.routeHelper = routeHelper;
         this.autoPublishComments = configuration.getBoolean(Configuration.COMMENTS_AUTO_PUBLISH,
                 Configuration.DEFAULT_COMMENTS_AUTO_PUBLISH);
     }
@@ -84,17 +78,17 @@ public class EdgeSetProperty implements ParameterizedHandler {
             User user,
             Authorizations authorizations
     ) throws Exception {
-        routeHelper.checkVisibilityParameter(visibilitySource, authorizations, user, resourceBundle);
-        routeHelper.checkRoutePath("edge", propertyName, request);
+        checkVisibilityParameter(visibilitySource, authorizations, user, resourceBundle);
+        checkRoutePath("edge", propertyName, request);
 
-        boolean isComment = routeHelper.isCommentProperty(propertyName);
+        boolean isComment = isCommentProperty(propertyName);
         boolean autoPublish = isComment && autoPublishComments;
         if (autoPublish) {
             workspaceId = null;
         }
 
         if (propertyKey == null) {
-            propertyKey = routeHelper.createPropertyKey(propertyName, graph);
+            propertyKey = createPropertyKey(propertyName, graph);
         }
 
         Edge edge = graph.getEdge(edgeId, authorizations);
